Rename DrawingToolRoot to Mark Rename the generic mark component so that its name matches the Mark model.
