--- packages/lib-classifier/src/plugins/drawingTools/components/DrawingToolRoot/index.js
@@ -1,1 +1,0 @@
-export { default } from './DrawingToolRoot'
