+++ packages/lib-classifier/src/plugins/drawingTools/components/Point/Point.stories.js
@@ -4,7 +4,7 @@ import zooTheme from '@zooniverse/grommet-theme'
 import React, { Component, createRef, forwardRef } from 'react'
 import sinon from 'sinon'
 import { PointTool } from '@plugins/drawingTools/models/tools'
-import { DrawingToolRoot }from '@plugins/drawingTools/components'
+import { Mark }from '@plugins/drawingTools/components'
 import Point from './Point'
 
 class DrawingStory extends Component {
@@ -13,12 +13,12 @@ class DrawingStory extends Component {
 
     return (
       <svg viewBox='0 0 300 400' height={300} width={400}>
-        <DrawingToolRoot
+        <Mark
           label='Point'
           mark={mark}
         >
           {children}
-        </DrawingToolRoot>
+        </Mark>
       </svg>
     )
   }
